+++ GHEtool/main_class.py
@@ -15,7 +15,7 @@ from scipy.signal import convolve
 from warnings import warn
 
 from GHEtool.VariableClasses import FluidData, PipeData, Borehole, GroundConstantTemperature
-from GHEtool.VariableClasses import CustomGFunction, load_custom_gfunction, GFunction, SizingSetup
+from GHEtool.VariableClasses import CustomGFunction, load_custom_gfunction, GFunction, CalculationSetup
 from GHEtool.VariableClasses.LoadData import *
 from GHEtool.VariableClasses.LoadData import _LoadData
 from GHEtool.VariableClasses.BaseClass import BaseClass
@@ -36,16 +36,16 @@ class Borefield(BaseClass):
 
     HOURLY_LOAD_ARRAY: np.ndarray = np.arange(0, 8761, UPM).astype(np.uint32)
 
-    __slots__ = 'H', 'H_init', 'borehole',\
-                'number_of_boreholes', '_borefield', 'cost_investment', \
-                'Tf_max', 'Tf_min', 'limiting_quadrant', \
-                'Tf', \
-                '_ground_data', '_borefield_load',\
-                'results_peak_heating', 'options_pygfunction',\
-                'results_peak_cooling', 'results_month_cooling', 'results_month_heating', 'Tb', 'THRESHOLD_WARNING_SHALLOW_FIELD', \
-                'gui', 'D', 'r_b', 'gfunction_calculation_object',\
-                'H_init', 'use_precalculated_data', '_sizing_setup', \
-                '_secundary_borefield_load', '_building_load', '_external_load'
+    __slots__ = 'H', 'borehole', \
+        'number_of_boreholes', '_borefield', 'cost_investment', \
+        'Tf_max', 'Tf_min', 'limiting_quadrant', \
+        'Tf', \
+        '_ground_data', '_borefield_load', \
+        'results_peak_heating', 'options_pygfunction', \
+        'results_peak_cooling', 'results_month_cooling', 'results_month_heating', 'Tb', 'THRESHOLD_WARNING_SHALLOW_FIELD', \
+        'gui', 'D', 'r_b', 'gfunction_calculation_object', \
+        '_sizing_setup', \
+        '_secundary_borefield_load', '_building_load', '_external_load'
 
     def __init__(self, peak_heating: np.ndarray | list = None,
                  peak_cooling: np.ndarray | list = None,
@@ -100,10 +100,12 @@ class Borefield(BaseClass):
 
         create the borefield object
 
-        >>> borefield = Borefield(peak_heating=peak_heating,
-        >>>                      peak_cooling=peak_cooling,
-        >>>                      baseload_heating=monthly_load_heating,
-        >>>                      baseload_cooling=monthly_load_cooling)
+        >>> borefield = Borefield()
+
+        set the load
+
+        >>> load = MonthlyGeothermalLoadAbsolute(monthly_load_heating, monthly_load_cooling, peak_heating, peak_cooling)
+        >>> borefield.load = load
 
         """
 
@@ -122,10 +124,6 @@ class Borefield(BaseClass):
         # m hereafter one needs to chance to fewer boreholes with more depth, because the calculations are no longer
         # that accurate.
         self.THRESHOLD_WARNING_SHALLOW_FIELD: int = 50
-
-        # setting this to False will make sure every g-value is calculated on the spot
-        # this will make everything way slower!
-        self.use_precalculated_data: bool = True
 
         self.custom_gfunction: CustomGFunction = custom_gfunction
         self.gfunction_calculation_object: GFunction = GFunction()
@@ -156,8 +154,7 @@ class Borefield(BaseClass):
         self.borehole = Borehole()
 
         # initiate different sizing
-        self._sizing_setup: SizingSetup = SizingSetup()
-        self.H_init: float = 100.
+        self._sizing_setup: CalculationSetup = CalculationSetup()
         self.sizing_setup()
 
         # check if the GHEtool is used by the gui i
@@ -231,7 +228,8 @@ class Borefield(BaseClass):
         """
         self.borefield = borefield
 
-    def create_rectangular_borefield(self, N_1: int, N_2: int, B_1: int, B_2: int, H: float, D: float = 1, r_b: float = 0.075):
+    def create_rectangular_borefield(self, N_1: int, N_2: int, B_1: int, B_2: int, H: float, D: float = 1,
+                                     r_b: float = 0.075):
         """
         This function creates a rectangular borefield.
         It calls the pygfunction module in the background.
@@ -379,7 +377,7 @@ class Borefield(BaseClass):
         self.custom_gfunction = load_custom_gfunction(location)
         ghe_logger.main_info("Custom g-function has been loaded.")
 
-    def set_investment_cost(self, investment_cost: list =None) -> None:
+    def set_investment_cost(self, investment_cost: list = None) -> None:
         """
         This function sets the investment cost. This is linear with respect to the total field length.
         If None, the default is set.
@@ -485,7 +483,7 @@ class Borefield(BaseClass):
         Load data: LoadData
         """
         return self._borefield_load
-    
+
     @load.setter
     def load(self, load: _LoadData) -> None:
         """
@@ -502,7 +500,7 @@ class Borefield(BaseClass):
         """
         self._borefield_load = load
         self._delete_calculated_temperatures()
-    
+
     @property
     def simulation_period(self) -> int:
         """
@@ -529,7 +527,7 @@ class Borefield(BaseClass):
         None
         """
         self._borefield_load.simulation_period = simulation_period
-        
+
     @property
     def Rb(self) -> float:
         """
@@ -718,6 +716,41 @@ class Borefield(BaseClass):
 
         return self.ground_data.calculate_Tg(H)
 
+    def _check_convergence(self, new_depth: float, old_depth: float, iter: int) -> bool:
+        """
+        This function checks, with the absolute and relative tolerance, if the depth is converged.
+        This is the case if both criteria are met. Raises runtime error if the max number of iterations is crossed.
+
+        Parameters
+        ----------
+        new_depth : float
+            Depth from the current interation
+        old_depth : float
+            Depth from the previous iteration
+        iter : int
+            Current number of iteration
+
+        Returns
+        -------
+        bool
+            True if the depth is converged
+
+        Raises
+        ------
+        RuntimeError
+            RuntimeError if the max number of iterations is crossed
+        """
+        if iter + 1 > self._sizing_setup.max_nb_of_iterations:
+            raise RuntimeError(f'The maximum number of iterations {self._sizing_setup.max_nb_of_iterations} is crossed.'
+                               f'There is no size convergence.')
+        if old_depth == 0:
+            return False
+        test_a_tol = abs(new_depth - old_depth) <= self._sizing_setup.atol if self._sizing_setup.atol != False else True
+        test_rtol = abs(
+            new_depth - old_depth) / old_depth <= self._sizing_setup.rtol if self._sizing_setup.rtol != False else True
+
+        return test_a_tol and test_rtol
+
     def _Ahmadfard(self, th: float, qh: float, qm: float, qa: float) -> float:
         """
         This function sizes the field based on the last year of operation, i.e. quadrants 2 and 4.
@@ -756,7 +789,8 @@ class Borefield(BaseClass):
         time = np.array([th, th + self.load.tm, self.load.ty + self.load.tm + th])
         # Iterates as long as there is no convergence
         # (convergence if difference between depth in iterations is smaller than THRESHOLD_BOREHOLE_DEPTH)
-        while abs(self.H - H_prev) >= Borefield.THRESHOLD_BOREHOLE_DEPTH:
+        i = 0
+        while not self._check_convergence(self.H, H_prev, i):
             # calculate the required g-function values
             gfunct_uniform_T = self.gfunction(time, max(1, self.H))
             # calculate the thermal resistances
@@ -768,6 +802,8 @@ class Borefield(BaseClass):
             # updating the depth values
             H_prev = self.H
             self.H = L / self.number_of_boreholes
+            i += 1
+
         return self.H
 
     def _Carcel(self, th: float, tcm: float, qh: float, qpm: float, qm: float) -> float:
@@ -797,7 +833,8 @@ class Borefield(BaseClass):
 
         # Iterates as long as there is no convergence
         # (convergence if difference between depth in iterations is smaller than THRESHOLD_BOREHOLE_DEPTH)
-        while abs(self.H - H_prev) >= Borefield.THRESHOLD_BOREHOLE_DEPTH:
+        i = 0
+        while not self._check_convergence(self.H, H_prev, i):
             # get the g-function values
             gfunc_uniform_T = self.gfunction(time_steps, max(1, self.H))
 
@@ -812,10 +849,10 @@ class Borefield(BaseClass):
             # updating the depth values
             H_prev = self.H
             self.H = L / self.number_of_boreholes
+            i += 1
         return self.H
 
-    def sizing_setup(self, H_init: float = None, use_constant_Rb: bool = None, quadrant_sizing: int = 0,
-                     L2_sizing: bool = None, L3_sizing: bool = None, L4_sizing: bool = None, sizing_setup: SizingSetup = None) -> None:
+    def sizing_setup(self, sizing_setup: CalculationSetup = None, use_constant_Rb: bool = None, **kwargs) -> None:
         """
         This function sets the options for the sizing function.
 
@@ -827,22 +864,14 @@ class Borefield(BaseClass):
 
         Parameters
         ----------
-        H_init : float
-            Initial depth of the borefield to start the iteration (m)
+        sizing_setup : CalculationSetup
+            An instance of the CalculationSetup class. When this argument differs from None, all the other parameters are
+            set based on this calculation_setup
         use_constant_Rb : bool
             True if a constant borehole equivalent resistance (Rb*) value should be used
-        quadrant_sizing : int
-            Differs from 0 when a sizing in a certain quadrant is desired.
-            Quadrants are developed by (Peere et al., 2021) [#PeereBS]_, [#PeereThesis]_
-        L2_sizing : bool
-            True if a sizing with the L2 method is needed
-        L3_sizing : bool
-            True if a sizing with the L3 method is needed
-        L4_sizing : bool
-            True if a sizing with the L4 method is needed
-        sizing_setup : SizingSetup
-            An instance of the SizingSetup class. When this argument differs from None, all the other parameters are
-            set based on this sizing_setup
+        kwargs
+            Dictionary with all the other options that can be set within GHEtool. For a complete list,
+            see the documentation in the CalculationSetup class.
 
         Returns
         -------
@@ -853,23 +882,20 @@ class Borefield(BaseClass):
         .. [#PeereBS] Peere, W., Picard, D., Cupeiro Figueroa, I., Boydens, W., and Helsen, L. (2021) Validated combined first and last year borefield sizing methodology. In Proceedings of International Building Simulation Conference 2021. Brugge (Belgium), 1-3 September 2021. https://doi.org/10.26868/25222708.2021.30180
         .. [#PeereThesis] Peere, W. (2020) Methode voor economische optimalisatie van geothermische verwarmings- en koelsystemen. Master thesis, Department of Mechanical Engineering, KU Leuven, Belgium.
         """
-        if H_init is not None:
-            self.H_init = H_init
 
         # if sizing_setup is not None, then the sizing setup is set directly
         if sizing_setup is not None:
             self._sizing_setup = sizing_setup
             return
 
-        self._sizing_setup = SizingSetup(quadrant_sizing=quadrant_sizing,
-                                         L2_sizing=L2_sizing,
-                                         L3_sizing=L3_sizing,
-                                         L4_sizing=L4_sizing)
+        self._sizing_setup = CalculationSetup()
+        self._sizing_setup.update_variables(**kwargs)
+
         if not use_constant_Rb is None:
             self.borehole.use_constant_Rb = use_constant_Rb
 
     def size(self, H_init: float = None, use_constant_Rb: bool = None, L2_sizing: bool = None,
-             L3_sizing: bool = None, L4_sizing: bool = None, quadrant_sizing: int = None) -> float:
+             L3_sizing: bool = None, L4_sizing: bool = None, quadrant_sizing: int = None, **kwargs) -> float:
         """
         This function sets the options for the sizing function.
 
@@ -887,15 +913,18 @@ class Borefield(BaseClass):
             Initial depth for the iteration. If None, the default H_init is chosen.
         use_constant_Rb : bool
             True if a constant borehole equivalent resistance (Rb*) value should be used
-        quadrant_sizing : int
-            Differs from 0 when a sizing in a certain quadrant is desired.
-            Quadrants are developed by (Peere et al., 2021) [#PeereBS]_, [#PeereThesis]_
         L2_sizing : bool
             True if a sizing with the L2 method is needed
         L3_sizing : bool
             True if a sizing with the L3 method is needed
         L4_sizing : bool
             True if a sizing with the L4 method is needed
+        quadrant_sizing : int
+            Differs from 0 when a sizing in a certain quadrant is desired.
+            Quadrants are developed by (Peere et al., 2021) [#PeereBS]_, [#PeereThesis]_
+        kwargs : dict
+            Dictionary with all the other options that can be set within GHEtool. For a complete list,
+            see the documentation in the CalculationSetup class.
 
         Returns
         -------
@@ -910,17 +939,16 @@ class Borefield(BaseClass):
         if not self.ground_data.check_values():
             raise ValueError("Please provide ground data.")
 
-        # check H_init
-        if H_init is None:
-            H_init = self.H_init
-
         # make backup of initial parameter states
         self._sizing_setup.make_backup()
         use_constant_Rb_backup = self.borehole.use_constant_Rb
 
         # run the sizing setup
-        self._sizing_setup.update_variables(L2_sizing=L2_sizing, L3_sizing=L3_sizing, L4_sizing=L4_sizing,
+        self._sizing_setup.update_variables(H_init=H_init, L2_sizing=L2_sizing, L3_sizing=L3_sizing,
+                                            L4_sizing=L4_sizing,
                                             quadrant_sizing=quadrant_sizing)
+        self._sizing_setup.update_variables(**kwargs)
+
         if not use_constant_Rb is None:
             self.borehole.use_constant_Rb = use_constant_Rb
 
@@ -1018,7 +1046,7 @@ class Borefield(BaseClass):
             raise ValueError(f'Quadrant {quadrant_sizing} does not exist.')
 
         # initiate with a given depth
-        self.H: float = H_init if H_init is not None else self.H_init
+        self.H: float = H_init if H_init is not None else self._sizing_setup.H_init
 
         def size_quadrant1():
             th, _, tcm, qh, qpm, qm = self.load._calculate_first_year_params(False)  # calculate parameters
@@ -1114,7 +1142,7 @@ class Borefield(BaseClass):
             raise ValueError(f'Quadrant {quadrant_sizing} does not exist.')
 
         # initiate with a given depth
-        self.H: float = H_init if H_init is not None else self.H_init
+        self.H: float = H_init if H_init is not None else self._sizing_setup.H_init
 
         if quadrant_sizing != 0:
             # size according to a specific quadrant
@@ -1173,14 +1201,16 @@ class Borefield(BaseClass):
             raise ValueError("There is no hourly resolution available!")
 
         # initiate with a given depth
-        self.H: float = H_init if H_init is not None else self.H_init
+        self.H: float = H_init if H_init is not None else self._sizing_setup.H_init
 
         if quadrant_sizing != 0:
             # size according to a specific quadrant
             self.H = self._size_based_on_temperature_profile(quadrant_sizing, hourly=True)
         else:
-            max_temp = self._size_based_on_temperature_profile(10, hourly=True) if np.any(self.load.hourly_cooling_load) else 0
-            min_temp = self._size_based_on_temperature_profile(20, hourly=True) if np.any(self.load.hourly_heating_load) else 0
+            max_temp = self._size_based_on_temperature_profile(10, hourly=True) if np.any(
+                self.load.hourly_cooling_load) else 0
+            min_temp = self._size_based_on_temperature_profile(20, hourly=True) if np.any(
+                self.load.hourly_heating_load) else 0
             self.H = self._select_size(max_temp, min_temp, True)
             if self.load.imbalance <= 0:
                 # extraction dominated, so quadrants 1 and 4 are relevant
@@ -1233,7 +1263,8 @@ class Borefield(BaseClass):
 
         # Iterates as long as there is no convergence
         # (convergence if difference between depth in iterations is smaller than THRESHOLD_BOREHOLE_DEPTH)
-        while abs(self.H - H_prev) >= Borefield.THRESHOLD_BOREHOLE_DEPTH:
+        i = 0
+        while not self._check_convergence(self.H, H_prev, i):
 
             if hourly:
                 # # calculate g-values
@@ -1317,19 +1348,23 @@ class Borefield(BaseClass):
             if quadrant == 1:
                 # maximum temperature
                 # convert back to required length
-                self.H = (np.max(self.results_peak_cooling[:8760 if hourly else 12]) - self._Tg()) / (self.Tf_max - self._Tg()) * H_prev
+                self.H = (np.max(self.results_peak_cooling[:8760 if hourly else 12]) - self._Tg()) / (
+                        self.Tf_max - self._Tg()) * H_prev
             elif quadrant == 2:
                 # maximum temperature
                 # convert back to required length
-                self.H = (np.max(self.results_peak_cooling[-8760 if hourly else -12:]) - self._Tg()) / (self.Tf_max - self._Tg()) * H_prev
+                self.H = (np.max(self.results_peak_cooling[-8760 if hourly else -12:]) - self._Tg()) / (
+                        self.Tf_max - self._Tg()) * H_prev
             elif quadrant == 3:
                 # minimum temperature
                 # convert back to required length
-                self.H = (np.min(self.results_peak_heating[:8760 if hourly else 12]) - self._Tg()) / (self.Tf_min - self._Tg()) * H_prev
+                self.H = (np.min(self.results_peak_heating[:8760 if hourly else 12]) - self._Tg()) / (
+                        self.Tf_min - self._Tg()) * H_prev
             elif quadrant == 4:
                 # minimum temperature
                 # convert back to required length
-                self.H = (np.min(self.results_peak_heating[-8760 if hourly else -12:]) - self._Tg()) / (self.Tf_min - self._Tg()) * H_prev
+                self.H = (np.min(self.results_peak_heating[-8760 if hourly else -12:]) - self._Tg()) / (
+                        self.Tf_min - self._Tg()) * H_prev
             elif quadrant == 10:
                 # over all years
                 # maximum temperature
@@ -1343,6 +1378,8 @@ class Borefield(BaseClass):
 
             if self.H < 0:
                 return 0
+
+        i += 1
 
         return self.H
 
@@ -1553,13 +1590,15 @@ class Borefield(BaseClass):
             ax.step(time_array, self.results_peak_cooling, 'b-', where="post", lw=1.5, label='Tf peak cooling')
             ax.step(time_array, self.results_peak_heating, 'r-', where="post", lw=1.5, label='Tf peak heating')
 
-            ax.step(time_array, self.results_month_cooling, color='b', linestyle="dashed", where="post", lw=1.5, label='Tf base cooling')
-            ax.step(time_array, self.results_month_heating, color='r', linestyle="dashed", where="post", lw=1.5, label='Tf base heating')
+            ax.step(time_array, self.results_month_cooling, color='b', linestyle="dashed", where="post", lw=1.5,
+                    label='Tf base cooling')
+            ax.step(time_array, self.results_month_heating, color='r', linestyle="dashed", where="post", lw=1.5,
+                    label='Tf base heating')
 
         # define temperature bounds
         ax.hlines(self.Tf_min, 0, self.simulation_period, colors='r', linestyles='dashed', label='', lw=1)
         ax.hlines(self.Tf_max, 0, self.simulation_period, colors='b', linestyles='dashed', label='', lw=1)
-        ax.set_xticks(range(0, self.simulation_period+1, 2))
+        ax.set_xticks(range(0, self.simulation_period + 1, 2))
 
         # Plot legend
         if legend:
@@ -1625,7 +1664,7 @@ class Borefield(BaseClass):
 
             # convolution to get the monthly results
             results = convolve(self.load.monthly_average_load_simulation_period * 1000,
-                               g_value_differences)[:12*self.simulation_period]
+                               g_value_differences)[:12 * self.simulation_period]
 
             # calculation the borehole wall temperature for every month i
             Tb = results / (2 * pi * self.ground_data.k_s) / (H * self.number_of_boreholes) + self._Tg(H)
@@ -1634,15 +1673,19 @@ class Borefield(BaseClass):
             # now the Tf will be calculated based on
             # Tf = Tb + Q * R_b
             results_month_cooling = Tb + self.load.baseload_cooling_power_simulation_period * 1000 \
-                              * (Rb / self.number_of_boreholes / H)
+                                    * (Rb / self.number_of_boreholes / H)
             results_month_heating = Tb - self.load.baseload_heating_power_simulation_period * 1000 \
-                              * (Rb / self.number_of_boreholes / H)
+                                    * (Rb / self.number_of_boreholes / H)
 
             # extra summation if the g-function value for the peak is included
-            results_peak_cooling = results_month_cooling + (self.load.peak_cooling_simulation_period - self.load.baseload_cooling_power_simulation_period) * 1000 \
-                                     * (g_value_peak_cooling / self.ground_data.k_s / 2 / pi + Rb) / self.number_of_boreholes / H
-            results_peak_heating = results_month_heating - (self.load.peak_heating_simulation_period - self.load.baseload_heating_power_simulation_period) * 1000 \
-                                   * (g_value_peak_heating / self.ground_data.k_s / 2 / pi + Rb) / self.number_of_boreholes / H
+            results_peak_cooling = results_month_cooling + (
+                    self.load.peak_cooling_simulation_period - self.load.baseload_cooling_power_simulation_period) * 1000 \
+                                   * (
+                                           g_value_peak_cooling / self.ground_data.k_s / 2 / pi + Rb) / self.number_of_boreholes / H
+            results_peak_heating = results_month_heating - (
+                    self.load.peak_heating_simulation_period - self.load.baseload_heating_power_simulation_period) * 1000 \
+                                   * (
+                                           g_value_peak_heating / self.ground_data.k_s / 2 / pi + Rb) / self.number_of_boreholes / H
 
             # save temperatures under variable
             self.results_peak_heating = results_peak_heating
@@ -1738,12 +1781,12 @@ class Borefield(BaseClass):
             """
             # set the correct depth of the borefield
             self._update_borefield_depth(H=H)
-            return self.gfunction_calculation_object.calculate(time_value, self.borefield, self.ground_data.alpha)
+            return self.gfunction_calculation_object.calculate(time_value, self.borefield, self.ground_data.alpha,
+                                                               interpolate=self._sizing_setup.interpolate_gfunctions)
 
         ## 1 bypass any possible precalculated g-functions
-
         # if calculate is False, then the gfunctions are calculated jit
-        if not self.use_precalculated_data:
+        if not self._sizing_setup.use_precalculated_dataset:
             return jit_gfunction_calculation()
 
         ## 2 use precalculated g-functions when available
@@ -1804,12 +1847,12 @@ class Borefield(BaseClass):
         -------
         Reynolds number : float
         """
-        u = self.borehole.fluid_data.mfr / self.borehole.pipe_data.number_of_pipes / self.borehole.fluid_data.rho /\
+        u = self.borehole.fluid_data.mfr / self.borehole.pipe_data.number_of_pipes / self.borehole.fluid_data.rho / \
             (pi * self.borehole.pipe_data.r_in ** 2)
         return self.borehole.fluid_data.rho * u * self.borehole.pipe_data.r_in * 2 / self.borehole.fluid_data.mu
 
-    def optimise_load_profile(self, building_load: HourlyGeothermalLoad, depth: float = None, SCOP: float = 10**6,
-                              SEER: float = 10**6, print_results: bool = False) -> None:
+    def optimise_load_profile(self, building_load: HourlyGeothermalLoad, depth: float = None, SCOP: float = 10 ** 6,
+                              SEER: float = 10 ** 6, print_results: bool = False) -> None:
         """
         This function optimises the load based on the given borefield and the given hourly load.
         (When the load is not geothermal, the SCOP and SEER are used to convert it to a geothermal load.)
@@ -1861,8 +1904,8 @@ class Borefield(BaseClass):
 
         # load hourly heating and cooling load and convert it to geothermal loads
         primary_geothermal_load = HourlyGeothermalLoad()
-        primary_geothermal_load.set_hourly_cooling(building_load.hourly_cooling_load.copy() * (1 + 1/SEER))
-        primary_geothermal_load.set_hourly_heating(building_load.hourly_heating_load.copy() * (1 - 1/SCOP))
+        primary_geothermal_load.set_hourly_cooling(building_load.hourly_cooling_load.copy() * (1 + 1 / SEER))
+        primary_geothermal_load.set_hourly_heating(building_load.hourly_heating_load.copy() * (1 - 1 / SCOP))
 
         # set geothermal load
         self.load = primary_geothermal_load
@@ -1914,8 +1957,8 @@ class Borefield(BaseClass):
 
         # calculate the resulting secundary hourly profile that can be put on the borefield
         self._secundary_borefield_load = HourlyGeothermalLoad()
-        self._secundary_borefield_load.set_hourly_cooling(self.load.hourly_cooling_load / (1 + 1/SEER))
-        self._secundary_borefield_load.set_hourly_heating(self.load.hourly_heating_load / (1 - 1/SCOP))
+        self._secundary_borefield_load.set_hourly_cooling(self.load.hourly_cooling_load / (1 + 1 / SEER))
+        self._secundary_borefield_load.set_hourly_heating(self.load.hourly_heating_load / (1 - 1 / SCOP))
 
         # set building load
         self._building_load = building_load
@@ -1933,7 +1976,8 @@ class Borefield(BaseClass):
 
         if print_results:
             # print results
-            print("The peak load heating is: ", f'{self._secundary_borefield_load.max_peak_heating:.0f}', "kW, leading to",
+            print("The peak load heating is: ", f'{self._secundary_borefield_load.max_peak_heating:.0f}',
+                  "kW, leading to",
                   f'{np.sum(self._secundary_borefield_load.hourly_heating_load):.0f}', "kWh of heating.")
             print("This is", f'{self._percentage_heating:.0f}',
                   "% of the total heating load.")
@@ -1941,7 +1985,8 @@ class Borefield(BaseClass):
                   "kWh of heating should come from another source, with a peak of",
                   f'{self._external_load.max_peak_heating:.0f}', "kW.")
             print("------------------------------------------")
-            print("The peak load cooling is: ", f'{self._secundary_borefield_load.max_peak_cooling:.0f}', "kW, leading to",
+            print("The peak load cooling is: ", f'{self._secundary_borefield_load.max_peak_cooling:.0f}',
+                  "kW, leading to",
                   f'{np.sum(self._secundary_borefield_load.hourly_cooling_load):.0f}', "kWh of cooling.")
             print("This is", f'{self._percentage_cooling:.0f}',
                   "% of the total cooling load.")
@@ -1962,7 +2007,7 @@ class Borefield(BaseClass):
         float
             Percentage of heating load that can be done geothermally.
         """
-        return np.sum(self._secundary_borefield_load.hourly_heating_load) /\
+        return np.sum(self._secundary_borefield_load.hourly_heating_load) / \
             np.sum(self._building_load.hourly_heating_load) * 100
 
     @property
@@ -1975,7 +2020,7 @@ class Borefield(BaseClass):
         float
             Percentage of cooling load that can be done geothermally.
         """
-        return np.sum(self._secundary_borefield_load.hourly_cooling_load) /\
+        return np.sum(self._secundary_borefield_load.hourly_cooling_load) / \
             np.sum(self._building_load.hourly_cooling_load) * 100
 
     def calculate_quadrant(self) -> int:
@@ -2100,8 +2145,12 @@ class Borefield(BaseClass):
         for i in range(self.borehole.pipe_data.number_of_pipes):
             circles_outer.append(plt.Circle(pos[i], self.borehole.pipe_data.r_out, color="black"))
             circles_inner.append(plt.Circle(pos[i], self.borehole.pipe_data.r_in, color="red"))
-            circles_outer.append(plt.Circle(pos[i + self.borehole.pipe_data.number_of_pipes], self.borehole.pipe_data.r_out, color="black"))
-            circles_inner.append(plt.Circle(pos[i + self.borehole.pipe_data.number_of_pipes], self.borehole.pipe_data.r_in, color="blue"))
+            circles_outer.append(
+                plt.Circle(pos[i + self.borehole.pipe_data.number_of_pipes], self.borehole.pipe_data.r_out,
+                           color="black"))
+            circles_inner.append(
+                plt.Circle(pos[i + self.borehole.pipe_data.number_of_pipes], self.borehole.pipe_data.r_in,
+                           color="blue"))
 
         # set visual settings for figure
         axes.set_aspect('equal')